Extract managers to separate files
