+++ src/managers/font_manager.rs
@@ -1,0 +1,28 @@
+use sdl2::ttf::{ Font, Sdl2TtfContext };
+use managers::resource_manager::{ ResourceLoader, ResourceManager };
+
+pub type FontManager<'l> = ResourceManager<'l, FontDetails, Font<'l, 'static>, Sdl2TtfContext>;
+
+// Information needed to load a Font
+#[derive(PartialEq, Eq, Hash)]
+pub struct FontDetails {
+    pub path: String,
+    pub size: u16,
+}
+
+impl<'a> From<&'a FontDetails> for FontDetails {
+    fn from(details: &'a FontDetails) -> FontDetails {
+        FontDetails {
+            path: details.path.clone(),
+            size: details.size,
+        }
+    }
+}
+
+// Font Context knows how to load Fonts
+impl<'l> ResourceLoader<'l, Font<'l, 'static>> for Sdl2TtfContext {
+    type Args = FontDetails;
+    fn load(&'l self, details: &FontDetails) -> Result<Font<'l, 'static>, String> {
+        self.load_font(&details.path, details.size)
+    }
+}
